revamps the player class to work with our new design
feature/fix-backend

This commit does a number of things. First and foremost it changes the player
class extensively to reflect our new design choices. Next, it adds 100% test
coverage for these changes. It also adds some functionality to the hand class
to make the new player functions work. It also makes changes to the acesUp class
so that only acesUp functions ever need to be called by the controller. This
modularization should hopefully help the frontend work.